Matrix rollover implemented
Generation will not rollover from one side of the matrix to another,
continuing on the other side. Note that this can cause some unusual
results in sufficiently small matrices in which growths out of a single
point can meet after rolling over. These growths terminate at the
meeting point due to neighbors already being in the closed set, which
can cause distinct lines in the heatmap.